refactor(api-v3): use integers in ScrweaponHudStats

+++ source/core/ScrWeaponHudStats.cs
@@ -5,19 +5,19 @@ namespace SHVDN
 	[StructLayout(LayoutKind.Explicit, Size = 0x28)] // total size 40 bytes
 	public struct ScrWeaponHudStats
     {
-		[FieldOffset(0x00)] public byte Damage;
-		[FieldOffset(0x08)] public byte Speed;
-		[FieldOffset(0x10)] public byte Capacity;
-		[FieldOffset(0x18)] public byte Accuracy;
-		[FieldOffset(0x20)] public byte Range;
+		[FieldOffset(0x00)] public int Damage;
+		[FieldOffset(0x08)] public int Speed;
+		[FieldOffset(0x10)] public int Capacity;
+		[FieldOffset(0x18)] public int Accuracy;
+		[FieldOffset(0x20)] public int Range;
 
 		public ScrWeaponHudStats(int damage, int speed, int capacity, int accuracy, int range)
 		{
-            Damage = (byte)damage;
-            Speed = (byte)speed;
-            Capacity = (byte)capacity;
-            Accuracy = (byte)accuracy;
-            Range = (byte)range;
+            Damage = damage;
+            Speed = speed;
+            Capacity = capacity;
+            Accuracy = accuracy;
+            Range = range;
         }
 	}
 }
